Handle HTTP status with ResponseEntity in the controller

+++ src/main/java/masterSpringMvc/user/api/UserApiController.java
@@ -3,6 +3,8 @@ package masterSpringMvc.user.api;
 import masterSpringMvc.user.User;
 import masterSpringMvc.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -24,17 +26,30 @@ public class UserApiController {
     }
 
     @RequestMapping(value = "/users", method = RequestMethod.POST)
-    public User createUser(@RequestBody User user) {
-        return userRepository.save(user);
+    public ResponseEntity<User> createUser(@RequestBody User user) {
+        HttpStatus status = HttpStatus.OK;
+        if (!userRepository.exists(user.getEmail())) {
+            status = HttpStatus.CREATED;
+        }
+        User saved = userRepository.save(user);
+        return new ResponseEntity<>(saved, status);
     }
 
     @RequestMapping(value = "/user/{email}", method = RequestMethod.PUT)
-    public User updateUser(@PathVariable String email, @RequestBody User user) {
-        return userRepository.save(email, user);
+    public ResponseEntity<User> updateUser(@PathVariable String email, @RequestBody User user) {
+        if (!userRepository.exists(user.getEmail())) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        User saved = userRepository.save(email, user);
+        return new ResponseEntity<>(saved, HttpStatus.CREATED);
     }
 
     @RequestMapping(value = "/user/{email}", method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable String email) {
+    public ResponseEntity<User> deleteUser(@PathVariable String email) {
+        if (!userRepository.exists(email)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         userRepository.delete(email);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
